Geocoding
I've made it so that we can look up locations and fetch the closest station.
This adds a dependency of `geopy`.
This ignores several edge cases. Namely, it assumes a flat earth with rectangular coordinates corresponding to the latitude and longitude. I am assuming these edge cases will not be noticeable.

Expected behavior:
```
user: !noaa tide 8722588
scubot: <a listing for station #8722588>

user: !noaa tide 32.8548228,-79.9614