Initial fetch from NOAA instead of geocoding API
I've found the NOAA website's listing for the latitude and longitude of their stations. This removes a great deal of strain from the geocoding API, which is good because our initial fetch was extremely taxing on the geocoder.

It will still be necessary to merge this branch with the branch that has database caching, as NOAA's response times result in the globe taking several hours to initialize.