Dependency Injection of database
The object structure prior to this change did not make it easy to isolate the `StationGlobe` class, whether for testing or prefetching the database cache. I have separated it using dependency injection patterns.

NOTE: This breaks the initialization on L150. Waiting for @suclearnub to provide more information on the NOAA.module_db object scope to fix that